pyproject.toml
[tool.poetry]
name = "djangoproject2"
version = "0.1.0"
description = ""
authors = ["Chousspb <chous_spb@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
